pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plant-disease-uv"
version = "0.1.0"
description = "Clasificación de enfermedades en cultivos con ViT + PlantVillage"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Joan Velásquez Vélez" }]

classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

# Dependencias principales (con versiones mínimas sensatas)
dependencies = [
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "transformers>=4.43.0",
  "datasets>=2.19.0",
  "evaluate>=0.4.0",
  "scikit-learn>=1.4.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "pyyaml>=6.0.1",
  "tqdm>=4.66.0",
  "matplotlib>=3.8.0",
  "onnx>=1.16.0",
  "onnxruntime>=1.18.0; platform_system != 'Darwin'",
  "streamlit>=1.36.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "black>=24.0.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0",
]

[project.scripts]
# CLIs útiles
plant-disease-train = "plant_disease.training.train:main"
plant-disease-predict = "plant_disease.inference.predict:main"
plant-disease-app = "plant_disease.apps.streamlit_app:main"

[project.urls]
Homepage = "https://example.com"
Repository = "https://github.com/<tu-usuario>/<tu-repo>"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
filterwarnings = ["ignore::UserWarning"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "UP"]
ignore = []
exclude = ["data", "artifacts", "notebooks"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true